fix: dedupe test fixture into conftest; correct stale whoami principal wording
Addresses PR #9 review:
- SonarCloud: hoist the duplicated 23-line auth fixture + imports into
  tests/conftest.py, removing it from all 8 per-noun CLI test files
  (kills the new-code duplication that failed the quality gate).
- Qodo bug: whoami no longer returns a principal, so drop 'active principal'
  from the whoami --help text and the 'principal payload' line in
  explain auth diagnostics.
